+++ src/manage/denyAccessRequest.test.ts
@@ -85,7 +85,7 @@ describe("denyAccessRequest", () => {
     await expect(
       denyAccessRequest("https://some.resource/owner", mockAccessRequestVc())
     ).rejects.toThrow(
-      "Cannot discover consent endpoint from [https://pod-provider.iri/resource/.well-known/solid]: the well-known document contains no value for properties [http://www.w3.org/ns/solid/terms#consent] or [http://inrupt.com/ns/ess#consentIssuer]."
+      "Cannot discover consent endpoint from [https://pod-provider.iri/resource/.well-known/solid]: the well-known document contains no value for properties [http://www.w3.org/ns/solid/terms#accessIssuer] or [http://inrupt.com/ns/ess#consentIssuer]."
     );
   });